requirements.txt
absl-py==2.1.0
alembic==1.13.1
astunparse==1.6.3
attrs==23.2.0
boto3==1.34.114
botocore==1.34.114
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
clearml==1.16.1
clearml-agent==1.8.1
colorlog==6.8.2
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
distlib==0.3.8
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.5.0
furl==2.1.3
gast==0.5.4
gitdb==4.0.11
github3.py==3.2.0
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.64.0
h5py==3.11.0
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
load-dotenv==0.1.0
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
optuna==3.6.1
orderedmultidict==1.0.1
packaging==24.0
pandas==2.2.2
pathlib2==2.3.7.post1
pillow==10.3.0
platformdirs==4.2.2
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
s3transfer==0.10.1
scikit-learn==1.5.0
scipy==1.13.0
seaborn==0.13.2
setuptools==70.0.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.30
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.17
uritemplate==4.1.1
urllib3==1.26.18
virtualenv==20.26.2
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
